+++ src/jsbind/lipsum_binding_js.cpp
@@ -86,7 +86,8 @@ EMSCRIPTEN_BINDINGS(lipsumcpp)
             .value("Cpp", lpsm::CodeLanguage::Cpp)
             .value("Python", lpsm::CodeLanguage::Python)
             .value("Rust", lpsm::CodeLanguage::Rust)
-            .value("C", lpsm::CodeLanguage::C);
+            .value("C", lpsm::CodeLanguage::C)
+            .value("JavaScript", lpsm::CodeLanguage::JavaScript);
 
     constant("LIPSUM_CPP_VERSION", std::string(LIPSUM_CPP_VERSION));
     constant("LIPSUM_CPP_VERSION_MAJOR", LIPSUM_CPP_VERSION_MAJOR);